Create GradingMarks.java
You are hired as an Associate Software Engineer in a reputed Multi-National Company (MNC).  Your company has received a software requirement from Shri Dharmasthala Manjunatheshwara College of Engineering and Technology, Dharwad, a software that computes and publishes the  examination results. 
You are assigned to work on Students’ Grading System project. As part of this project, you are  asked by your team lead to write a GUI based Java program to compute the grade obtained by a student in a  single course. A sample GUI is given below